Fix GPX where "number" is decimal

+++ gpx/gpx10.go
@@ -209,7 +209,7 @@ type gpx10GpxRte struct {
 	Src     string   `xml:"src,omitempty"`
 	// TODO
 	//Links       []Link   `xml:"link"`
-	Number int              `xml:"number,omitempty"`
+	Number float64          `xml:"number,omitempty"`
 	Type   string           `xml:"type,omitempty"`
 	Points []*gpx10GpxPoint `xml:"rtept"`
 }
@@ -228,7 +228,7 @@ type gpx10GpxTrk struct {
 	Src     string   `xml:"src,omitempty"`
 	// TODO
 	//Links    []Link   `xml:"link"`
-	Number   int               `xml:"number,omitempty"`
+	Number   float64           `xml:"number,omitempty"`
 	Type     string            `xml:"type,omitempty"`
 	Segments []*gpx10GpxTrkSeg `xml:"trkseg,omitempty"`
 }
